Add support for error middleware
