+++ runtime/src/server/middleware/index.ts
@@ -1,22 +1,29 @@
 import fs from 'fs';
 import path from 'path';
 import mime from 'mime/lite';
-import { Handler, SapperRequest, SapperResponse, build_dir, dev, manifest } from '@sapper/internal/manifest-server';
+import { SapperRequest, SapperResponse, SapperNext, SapperHandler, SapperErrorHandler, build_dir, dev, manifest } from '@sapper/internal/manifest-server';
+import { get_page_renderer } from './get_page_renderer';
 import { get_server_route_handler } from './get_server_route_handler';
 import { get_page_handler } from './get_page_handler';
+import { get_error_handler } from './get_error_handler';
 
 type IgnoreValue = IgnoreValue[] | RegExp | ((uri: string) => boolean) | string;
 
 export default function middleware(opts: {
 	session?: (req: SapperRequest, res: SapperResponse) => any,
-	ignore?: IgnoreValue
+	ignore?: IgnoreValue,
+	catchErrors?: boolean
 } = {}) {
 	const { session, ignore } = opts;
 
 	let emitted_basepath = false;
 
+	const page_renderer = get_page_renderer(manifest, session || noop);
+	const page_handler = get_page_handler(manifest, page_renderer);
+	const error_handler = get_error_handler(manifest, page_renderer);
+
 	return compose_handlers(ignore, [
-		(req: SapperRequest, res: SapperResponse, next: () => void) => {
+		(req: SapperRequest, res: SapperResponse, next: SapperNext) => {
 			if (req.baseUrl === undefined) {
 				let originalUrl = req.originalUrl || req.url;
 				if (req.url === '/' && originalUrl[originalUrl.length - 1] !== '/') {
@@ -62,28 +69,37 @@ export default function middleware(opts: {
 
 		get_server_route_handler(manifest.server_routes),
 
-		get_page_handler(manifest, session || noop)
-	].filter(Boolean));
+		page_handler
+	].filter(Boolean), error_handler);
 }
 
-export function compose_handlers(ignore: IgnoreValue, handlers: Handler[]): Handler {
+export function compose_handlers(ignore: IgnoreValue, handlers: SapperHandler[], error_handler: SapperErrorHandler): SapperHandler {
 	const total = handlers.length;
 
-	function nth_handler(n: number, req: SapperRequest, res: SapperResponse, next: () => void) {
+	function nth_handler(n: number, req: SapperRequest, res: SapperResponse, next: SapperNext, error_next: SapperNext) {
 		if (n >= total) {
 			return next();
 		}
 
-		handlers[n](req, res, () => nth_handler(n+1, req, res, next));
+		const handler = handlers[n];
+		const handler_next: SapperNext = (err) => {
+			if (err) {
+				error_next(err);
+			} else {
+				nth_handler(n+1, req, res, next, error_next);
+			}
+		};
+
+		handler(req, res, handler_next);
 	}
 
 	return !ignore
-		? (req, res, next) => nth_handler(0, req, res, next)
+		? (req, res, next) => nth_handler(0, req, res, next, (err) => error_handler(err, req, res, next))
 		: (req, res, next) => {
 			if (should_ignore(req.path, ignore)) {
 				next();
 			} else {
-				nth_handler(0, req, res, next);
+				nth_handler(0, req, res, next, (err) => error_handler(err, req, res, next));
 			}
 		};
 }
@@ -99,7 +115,7 @@ export function serve({ prefix, pathname, cache_control }: {
 	prefix?: string,
 	pathname?: string,
 	cache_control: string
-}) {
+}): SapperHandler {
 	const filter = pathname
 		? (req: SapperRequest) => req.path === pathname
 		: (req: SapperRequest) => req.path.startsWith(prefix);
@@ -110,7 +126,7 @@ export function serve({ prefix, pathname, cache_control }: {
 		? (file: string) => fs.readFileSync(path.join(build_dir, file))
 		: (file: string) => (cache.has(file) ? cache : cache.set(file, fs.readFileSync(path.join(build_dir, file)))).get(file);
 
-	return (req: SapperRequest, res: SapperResponse, next: () => void) => {
+	return (req: SapperRequest, res: SapperResponse, next: SapperNext) => {
 		if (filter(req)) {
 			const type = mime.getType(req.path);
 
@@ -125,10 +141,7 @@ export function serve({ prefix, pathname, cache_control }: {
 				if (err.code === 'ENOENT') {
 					next();
 				} else {
-					console.error(err);
-
-					res.statusCode = 500;
-					res.end('an error occurred while reading a static file from disk');
+					next(err);
 				}
 			}
 		} else {
